Fix audio-path bugs, add typed Effect Graph port specs, restyle recording UI
Bug fixes:
- Use integer/f64 bin math in waveform min/max builders and speed offline
  processing; f32 accumulators lose sample-index precision past 2^24
  samples (~6 min at 48 kHz), skewing waveforms on long files
- Replace delete-then-rename file overwrites with unique sibling temp
  names and a parking rename, closing the crash window that could lose
  the original file and the fixed temp-name collision between
  concurrent operations in the same folder
- Apply loop crossfade once: blend styles crossfade in the tail window
  only and the wrap now skips the consumed head region; previously the
  seam was blended twice and jumped right after wrapping
- Mark header-only peak values (first 0.25 s estimate) with
  peak_db_estimate and render them as "~x.x" with a tooltip
- Round in f32_to_i16 instead of truncating, matching the WAV writer
- Remove dead SharedAudio state (file_gain, swap_* crossfade fields)
- Simplify redundant undo/redo hotkey boolean expression

Effect Graph type system:
- Add EffectGraphNodeSpec/EffectGraphPortSpec: a static per-kind table
  of display name, palette category, and typed ports (direction +
  EffectGraphPortType) as the single source of truth; palette, pin
  labels, and port lookups all derive from it
- Connect validation resolves both ends through the spec, checks port
  data-type compatibility, and no-ops on duplicate edges instead of
  churning the undo stack
- Make validate_effect_graph_document exhaustive (no catch-all) so new
  node kinds fail to compile until handled; replace the Mixed combine
  unreachable!() with a proper runtime error
- Serialized port-id strings are unchanged; a round-trip test guards
  spec/serde compatibility for every node kind

Recording UI:
- Card-based layout with dB-scaled level meters (colour zones, ticks,
  1.5 s peak hold), 96 px waveform, large monospace elapsed time with
  tenths while recording
- Painter-drawn record/pause toggle, sized transport buttons with
  tooltips, and a confirmation modal before discarding a take
- "Save As…" copies the finished take to a user-chosen WAV path
- cpal stream errors now reach the UI as RecordingState::Error (partial
  take is still finalized) and dropped capture buffers show an input
  overrun warning
- Disable the unimplemented "System + Mic" source (it recorded mic only)
  and hide system sources on non-Windows

+++ src/app/render/waveform_pyramid.rs
@@ -319,8 +319,9 @@ pub fn build_visible_minmax(samples: &[f32], bins: usize, out: &mut Vec<Peak>) {
     }
     out.reserve(bins.min(samples.len()));
     let len = samples.len();
-    let mut pos = 0.0f32;
-    let step = (len as f32 / bins as f32).max(1.0);
+    // f64: f32 mantissa cannot represent sample indices above 2^24 (~6 min at 48 kHz).
+    let mut pos = 0.0f64;
+    let step = (len as f64 / bins as f64).max(1.0);
     while (pos as usize) < len {
         let start = pos.floor() as usize;
         let mut end = (pos + step).floor() as usize;
@@ -367,8 +368,8 @@ pub fn build_mixdown_minmax_visible(
     let visible_len = end - start;
     out.reserve(bins.min(visible_len));
     let inv_channels = 1.0 / channels.len().max(1) as f32;
-    let mut pos = 0.0f32;
-    let step = (visible_len as f32 / bins as f32).max(1.0);
+    let mut pos = 0.0f64;
+    let step = (visible_len as f64 / bins as f64).max(1.0);
     while (pos as usize) < visible_len {
         let i0 = start + pos.floor() as usize;
         let mut i1 = start + (pos + step).floor() as usize;
